+++ sdk/python/pulumi_azure_native/dbformysql/v20210501/__init__.py
@@ -6,8 +6,10 @@ from ... import _utilities
 import typing
 # Export this package's modules as members:
 from ._enums import *
+from .configuration import *
 from .database import *
 from .firewall_rule import *
+from .get_configuration import *
 from .get_database import *
 from .get_firewall_rule import *
 from .get_get_private_dns_zone_suffix_execute import *